+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.32.0:
+* add `iconWidget` in `MarkerIcon` to show dynamic widget
+* support show interestPoint for drawing road manually
+* add deleteOldRoad in drawRoadManually to prevent delete last road if it not needed
 ## 0.31.0:
 * add drawMultiRoad in iOS side
 * fix request permission when map initialized on user location